refactor: station class화
