+++ src/main/kotlin/subway/app/MenuApplication.kt
@@ -3,12 +3,18 @@ package subway.app
 import subway.domain.Line
 import subway.domain.Section
 import subway.domain.Station
+import subway.service.DeleteSectionService
+import subway.service.DeleteStationService
+import subway.service.RegisterSectionService
+import subway.service.RegisterStationService
 import subway.view.getDownwardNameOfSectionToDelete
 import subway.view.getDownwardNameOfSectionToRegister
 import subway.view.getLineNameOfSectionToDelete
 import subway.view.getLineNameOfSectionToRegister
 import subway.view.getSectionDistance
 import subway.view.getSectionTime
+import subway.view.getStationNameToDelete
+import subway.view.getStationNameToRegister
 import subway.view.getUpwardNameOfSectionToDelete
 import subway.view.getUpwardNameOfSectionToRegister
 import subway.view.infoMessage
@@ -22,7 +28,9 @@ import subway.view.showCheckPath
 import subway.view.showMainPage
 import subway.view.showWholeTrack
 import subway.view.succeedDeleteSection
+import subway.view.succeedDeleteStation
 import subway.view.succeedRegisterSection
+import subway.view.succeedRegisterStation
 
 const val MENU_ONE = 1
 const val MENU_TWO = 2
@@ -58,8 +66,18 @@ fun adminStation() {
     if (select == BACK) return
 
     when (select.toInt()) {
-        MENU_ONE -> registerStation()
-        MENU_TWO -> deleteStation()
+        MENU_ONE -> {
+            val name = getStationNameToRegister()
+            RegisterStationService(Station(name)).register()
+            infoMessage()
+            succeedRegisterStation()
+        }
+        MENU_TWO -> {
+            val name = getStationNameToDelete()
+            DeleteStationService(name).delete()
+            infoMessage()
+            succeedDeleteStation()
+        }
         MENU_THREE -> showAllStations()
     }
 }
@@ -84,7 +102,7 @@ fun adminSection() {
 
     when (select.toInt()) {
         MENU_ONE -> {
-            RegisterSection(
+            RegisterSectionService(
                 Section(
                     line = Line(getLineNameOfSectionToRegister()),
                     upwardStation = Station(getUpwardNameOfSectionToRegister()),
@@ -97,7 +115,7 @@ fun adminSection() {
             succeedRegisterSection()
         }
         MENU_TWO -> {
-            DeleteSection(
+            DeleteSectionService(
                 line = Line(getLineNameOfSectionToDelete()),
                 upwardStation = Station(getUpwardNameOfSectionToDelete()),
                 downwardStation = Station(getDownwardNameOfSectionToDelete())
