+++ src/main/kotlin/subway/service/DeleteSectionService.kt
@@ -1,4 +1,4 @@
-package subway.app
+package subway.service
 
 import subway.domain.Line
 import subway.domain.Section
@@ -7,7 +7,7 @@ import subway.repository.LineRepository
 import subway.repository.SectionRepository
 import subway.repository.SectionRepository.existsByUpward
 
-class DeleteSection(
+class DeleteSectionService(
     val line: Line,
     val upwardStation: Station,
     val downwardStation: Station,
