+++ src/main/kotlin/subway/service/RegisterStationService.kt
@@ -1,0 +1,12 @@
+package subway.service
+
+import subway.domain.Station
+import subway.repository.StationRepository
+
+class RegisterStationService(val station: Station) {
+    init {
+        require(station.validStationToRegister())
+    }
+
+    fun register() = StationRepository.add(station)
+}
